Factor out FinishJobParams from the mock services
Can't factor is out of module runner for now since module runner has 0
tests, sigh
